refactor: moved sandbox resource creation logic into separate service
this makes sandbox service lighter, it's dependencies mockable and hence making it easier to test
