+++ src/Sepes.Infrastructure/Interface/IHasCurrentPhase.cs
@@ -1,0 +1,9 @@
+﻿using Sepes.Infrastructure.Model;
+
+namespace Sepes.Infrastructure.Interface
+{
+    public interface IHasCurrentPhase
+    {
+        SandboxPhase CurrentPhase { get; set; }
+    }
+}
