+++ src/Sepes.Infrastructure/Service/Interface/ISandboxService.cs
@@ -15,10 +15,7 @@ namespace Sepes.Infrastructure.Service.Interface
         Task<SandboxDetailsDto> CreateAsync(int studyId, SandboxCreateDto newSandbox);
         Task DeleteAsync(int sandboxId);
 
-        Task<List<SandboxResourceLightDto>> GetSandboxResources(int sandboxId);
-
-        Task<SandboxResourceLightDto> RetryLastOperation(int resourceId);
-        Task ReScheduleSandboxCreation(int sandboxId);
+      
 
         Task MoveToNextPhase(int sandboxId);
     }
