+++ src/Sepes.Infrastructure/Service/SandboxService.cs
@@ -3,10 +3,8 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Sepes.Infrastructure.Constants;
-using Sepes.Infrastructure.Constants.CloudResource;
 using Sepes.Infrastructure.Dto;
 using Sepes.Infrastructure.Dto.Sandbox;
-using Sepes.Infrastructure.Exceptions;
 using Sepes.Infrastructure.Interface;
 using Sepes.Infrastructure.Model;
 using Sepes.Infrastructure.Model.Context;
@@ -20,38 +18,27 @@ using System.Threading.Tasks;
 
 namespace Sepes.Infrastructure.Service
 {
-    public class SandboxService : ISandboxService
+    public class SandboxService : SandboxServiceBase, ISandboxService
     {
-        readonly IConfiguration _config;
-        readonly SepesDbContext _db;
-        readonly IMapper _mapper;
-        readonly ILogger _logger;
-        readonly IUserService _userService;
         readonly IRequestIdService _requestIdService;
         readonly IStudyService _studyService;
-        readonly ISandboxResourceService _sandboxResourceService;
-        readonly IProvisioningQueueService _provisioningQueueService;
+        readonly ISandboxCloudResourceService _sandboxCloudResourceService;     
 
 
         public SandboxService(IConfiguration config, SepesDbContext db, IMapper mapper, ILogger<SandboxService> logger,
-            IUserService userService, IRequestIdService requestIdService, IStudyService studyService, ISandboxResourceService sandboxResourceService, IProvisioningQueueService provisioningQueueService)
+            IUserService userService,
+            IRequestIdService requestIdService, IStudyService studyService, ISandboxCloudResourceService sandboxCloudResourceService)
+            :base (config, db, mapper, logger, userService)
         {
-            _db = db;
-            _mapper = mapper;
-            _logger = logger;
-            _userService = userService;
+        
             _requestIdService = requestIdService;
-            _studyService = studyService;
-            _sandboxResourceService = sandboxResourceService;
-            _provisioningQueueService = provisioningQueueService;
-            _config = config;
+            _studyService = studyService;       
+            _sandboxCloudResourceService = sandboxCloudResourceService;
         }
 
         public async Task<SandboxDto> GetAsync(int sandboxId, UserOperation userOperation)
         {
-            var sandboxFromDb = await GetOrThrowAsync(sandboxId, userOperation, false);
-            var sandboxDto = _mapper.Map<SandboxDto>(sandboxFromDb);
-            return sandboxDto;
+            return await GetDtoAsync(sandboxId, userOperation);
         }
 
         public async Task<SandboxDetailsDto> GetSandboxDetailsAsync(int sandboxId)
@@ -102,13 +89,16 @@ namespace Sepes.Infrastructure.Service
 
             try
             {
+                var user = await _userService.GetCurrentUserAsync();
+
                 createdSandbox = _mapper.Map<Sandbox>(sandboxCreateDto);
 
-                var user = await _userService.GetCurrentUserAsync();
-                createdSandbox.CreatedBy = user.UserName;
-                createdSandbox.TechnicalContactName = user.FullName;
-                createdSandbox.TechnicalContactEmail = user.EmailAddress;
+                InitiatePhaseHistory(createdSandbox, user);
 
+                createdSandbox.CreatedBy = user.UserName;               
+                createdSandbox.TechnicalContactName = user.FullName;
+                createdSandbox.TechnicalContactEmail = user.EmailAddress;               
+                
                 study.Sandboxes.Add(createdSandbox);
                 await _db.SaveChangesAsync();
 
@@ -119,14 +109,14 @@ namespace Sepes.Infrastructure.Service
                     var studyDto = await _studyService.GetStudyDtoByIdAsync(studyId, UserOperation.Study_Crud_Sandbox);
                     var sandboxDto = await GetAsync(createdSandbox.Id, UserOperation.Study_Crud_Sandbox);
 
-                    var tags = AzureResourceTagsFactory.SandboxResourceTags(_config, study, createdSandbox);
+                    var tags = AzureResourceTagsFactory.SandboxResourceTags(_configuration, study, createdSandbox);
 
                     var region = RegionStringConverter.Convert(sandboxCreateDto.Region);
 
                     //This objects gets passed around
                     var creationAndSchedulingDto = new SandboxResourceCreationAndSchedulingDto() { SandboxId = createdSandbox.Id, StudyName = studyDto.Name, SandboxName = sandboxDto.Name, Region = region, Tags = tags, BatchId = Guid.NewGuid().ToString() };
 
-                    await CreateBasicSandboxResourcesAsync(creationAndSchedulingDto);
+                    await _sandboxCloudResourceService.CreateBasicSandboxResourcesAsync(creationAndSchedulingDto);
                 }
                 catch (Exception ex)
                 {
@@ -142,152 +132,12 @@ namespace Sepes.Infrastructure.Service
             {
                 throw new Exception($"Sandbox creation failed: {ex.Message}", ex);
             }
-        }
-
-        async Task<Sandbox> GetOrThrowAsync(int sandboxId, UserOperation userOperation, bool withIncludes)
+        } 
+        
+        public void InitiatePhaseHistory(Sandbox sandbox, UserDto currentUser)
         {
-            var sandbox = await SandboxSingularQueries.GetSandboxByIdCheckAccessOrThrow(_db, _userService, sandboxId, userOperation, withIncludes);
-            return sandbox;
-        }
-
-        async Task<SandboxResourceCreationAndSchedulingDto> CreateBasicSandboxResourcesAsync(SandboxResourceCreationAndSchedulingDto dto)
-        {
-            _logger.LogInformation($"Creating basic sandbox resources for sandbox: {dto.SandboxName}. First creating Resource Group, other resources are created by worker");
-
-            try
-            {
-                await _sandboxResourceService.CreateSandboxResourceGroup(dto);
-
-                _logger.LogInformation($"Done creating Resource Group for sandbox: {dto.SandboxName}. Scheduling creation of other resources");
-
-                var queueParentItem = new ProvisioningQueueParentDto
-                {
-                    SandboxId = dto.SandboxId,
-                    Description = $"Create basic resources for Sandbox: {dto.SandboxId}"
-                };
-
-                await ScheduleCreationOfDiagStorageAccount(dto, queueParentItem);
-                await ScheduleCreationOfNetworkSecurityGroup(dto, queueParentItem);
-                await ScheduleCreationOfVirtualNetwork(dto, queueParentItem);
-                await ScheduleCreationOfBastion(dto, queueParentItem);
-
-                await _provisioningQueueService.SendMessageAsync(queueParentItem);
-
-                _logger.LogInformation($"Done ordering creation of basic resources for sandbox: {dto.SandboxName}");
-            }
-            catch (Exception ex)
-            {
-                throw new Exception($"Unable to create basic sandbox resources.", ex);
-            }
-
-            return dto;
-        }
-
-        async Task ScheduleCreationOfDiagStorageAccount(SandboxResourceCreationAndSchedulingDto dto, ProvisioningQueueParentDto queueParentItem)
-        {
-            var resourceName = AzureResourceNameUtil.DiagnosticsStorageAccount(dto.StudyName, dto.SandboxName);
-            var resourceGroupCreateOperation = dto.ResourceGroup.Operations.FirstOrDefault().Id;
-            var resourceEntry = await CreateResource(dto, queueParentItem, AzureResourceType.StorageAccount, sandboxControlled: true, resourceName: resourceName, dependsOn: resourceGroupCreateOperation);
-            dto.DiagnosticsStorage = resourceEntry;
-
-        }
-
-        async Task ScheduleCreationOfNetworkSecurityGroup(SandboxResourceCreationAndSchedulingDto dto, ProvisioningQueueParentDto queueParentItem)
-        {
-            var nsgName = AzureResourceNameUtil.NetworkSecGroupSubnet(dto.StudyName, dto.SandboxName);
-            var diagStorageAccountCreateOperation = dto.DiagnosticsStorage.Operations.FirstOrDefault().Id;
-            var resourceEntry = await CreateResource(dto, queueParentItem, AzureResourceType.NetworkSecurityGroup, sandboxControlled: true, resourceName: nsgName, dependsOn: diagStorageAccountCreateOperation);
-            dto.NetworkSecurityGroup = resourceEntry;
-        }
-
-        async Task ScheduleCreationOfVirtualNetwork(SandboxResourceCreationAndSchedulingDto dto, ProvisioningQueueParentDto queueParentItem)
-        {
-            //TODO: Add special network rules to resource
-            var networkName = AzureResourceNameUtil.VNet(dto.StudyName, dto.SandboxName);
-            var sandboxSubnetName = AzureResourceNameUtil.SubNet(dto.StudyName, dto.SandboxName);
-
-            var networkSettings = new NetworkSettingsDto() { SandboxSubnetName = sandboxSubnetName };
-            var networkSettingsString = SandboxResourceConfigStringSerializer.Serialize(networkSettings);
-
-            var nsgCreateOperation = dto.NetworkSecurityGroup.Operations.FirstOrDefault().Id;
-
-            var resourceEntry = await CreateResource(dto, queueParentItem, AzureResourceType.VirtualNetwork, sandboxControlled: true, resourceName: networkName, configString: networkSettingsString, dependsOn: nsgCreateOperation);
-            dto.Network = resourceEntry;
-        }
-
-        async Task ScheduleCreationOfBastion(SandboxResourceCreationAndSchedulingDto dto, ProvisioningQueueParentDto queueParentItem, string configString = null)
-        {
-            var vNetCreateOperation = dto.Network.Operations.FirstOrDefault().Id;
-
-            var bastionName = AzureResourceNameUtil.Bastion(dto.StudyName, dto.SandboxName);
-
-            var resourceEntry = await CreateResource(dto, queueParentItem, AzureResourceType.Bastion, sandboxControlled: true, resourceName: bastionName, configString: configString, dependsOn: vNetCreateOperation);
-            dto.Bastion = resourceEntry;
-        }
-
-        async Task<SandboxResourceDto> CreateResource(SandboxResourceCreationAndSchedulingDto dto, ProvisioningQueueParentDto queueParentItem, string resourceType, bool sandboxControlled = true, string resourceName = AzureResourceNameUtil.AZURE_RESOURCE_INITIAL_ID_OR_NAME, string configString = null, int dependsOn = 0)
-        {
-            var resourceEntry = await _sandboxResourceService.Create(dto, resourceType, sandboxControlled: sandboxControlled, resourceName: resourceName, configString: configString, dependsOn: dependsOn);
-            queueParentItem.Children.Add(new ProvisioningQueueChildDto() { SandboxResourceOperationId = resourceEntry.Operations.FirstOrDefault().Id });
-
-            return resourceEntry;
-        }
-
-        public async Task<SandboxResourceLightDto> RetryLastOperation(int resourceId)
-        {
-            var resource = await _sandboxResourceService.GetByIdAsync(resourceId);
-
-            var sandboxFromDb = await GetOrThrowAsync(resource.SandboxId, UserOperation.Study_Crud_Sandbox, true);
-
-            if (resource.ResourceType != AzureResourceType.VirtualMachine)
-            {
-                throw new ArgumentException("Retry is only supported for Virtual Machines");
-            }
-
-            var relevantOperation = resource.Operations.OrderByDescending(o => o.Created).FirstOrDefault();
-
-            if (relevantOperation == null)
-            {
-                throw new NullReferenceException(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxFromDb.Id, "Could not locate ANY database entry for VM", resourceId));
-            }
-            else if (String.IsNullOrWhiteSpace(relevantOperation.Status) || relevantOperation.Status == CloudResourceOperationState.NEW || relevantOperation.Status == CloudResourceOperationState.IN_PROGRESS || relevantOperation.Status == CloudResourceOperationState.FAILED || relevantOperation.Status == CloudResourceOperationState.DONE_SUCCESSFUL)
-            {
-                _logger.LogInformation(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxFromDb.Id, $"Increasing MAX try count", resourceId));
-
-                relevantOperation.MaxTryCount += CloudResourceConstants.RESOURCE_MAX_TRY_COUNT; //Increase max try count  
-
-                _logger.LogInformation(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxFromDb.Id, $"Re-queing item. Previous status was {relevantOperation.Status}", resourceId));
-
-                var queueParentItem = new ProvisioningQueueParentDto();
-                queueParentItem.SandboxId = sandboxFromDb.Id;
-                queueParentItem.Description = $"{relevantOperation} (re-scheduled)";
-
-                await _db.SaveChangesAsync();
-                queueParentItem.Children.Add(new ProvisioningQueueChildDto() { SandboxResourceOperationId = relevantOperation.Id });
-                await _provisioningQueueService.SendMessageAsync(queueParentItem);
-            }
-            else
-            {
-                throw new Exception(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxFromDb.Id, $"Could not locate RELEVANT database entry for ResourceGroupOperation", resourceId));
-            }
-
-            return _mapper.Map<SandboxResourceLightDto>(resource);
-        }
-
-        public async Task<List<SandboxResourceLightDto>> GetSandboxResources(int sandboxId)
-        {
-            var sandboxFromDb = await GetOrThrowAsync(sandboxId, UserOperation.Study_Read, true);
-
-            //Filter out deleted resources
-            var resourcesFiltered = sandboxFromDb.Resources
-                .Where(r => !r.Deleted.HasValue
-                || (r.Deleted.HasValue && r.Operations.Where(o => o.OperationType == CloudResourceOperationType.DELETE && o.Status == CloudResourceOperationState.DONE_SUCCESSFUL).Any() == false)
-
-                ).ToList();
-
-            var resourcesMapped = _mapper.Map<List<SandboxResourceLightDto>>(resourcesFiltered);
-
-            return resourcesMapped;
+            sandbox.PhaseHistory = new List<SandboxPhaseHistory>();
+            sandbox.PhaseHistory.Add(new SandboxPhaseHistory { Counter = 0, Phase = SandboxPhase.Open, CreatedBy = currentUser.UserName });
         }
 
         public async Task MoveToNextPhase(int sandboxId)
@@ -325,12 +175,7 @@ namespace Sepes.Infrastructure.Service
         {
             _logger.LogWarning(SepesEventId.SandboxDelete, "Sandbox {0}: Starting", sandboxId);
 
-            var sandboxFromDb = await GetOrThrowAsync(sandboxId, UserOperation.Study_Crud_Sandbox, true);
-
-            if (sandboxFromDb == null)
-            {
-                throw NotFoundException.CreateForEntity("Sandbox", sandboxId);
-            }
+            var sandboxFromDb = await GetOrThrowAsync(sandboxId, UserOperation.Study_Crud_Sandbox, true);         
 
             int studyId = sandboxFromDb.StudyId;
 
@@ -343,156 +188,11 @@ namespace Sepes.Infrastructure.Service
             sandboxFromDb.DeletedAt = DateTime.UtcNow;
             sandboxFromDb.DeletedBy = user.UserName;
 
-            SandboxResource sandboxResourceGroup = null;
-
-            if (sandboxFromDb.Resources.Count > 0)
-            {
-                //Mark all resources as deleted
-                foreach (var curResource in sandboxFromDb.Resources)
-                {
-                    if (curResource.ResourceType == AzureResourceType.ResourceGroup)
-                    {
-                        sandboxResourceGroup = curResource;
-                    }
-
-                    curResource.Deleted = DateTime.UtcNow;
-                    curResource.DeletedBy = user.UserName;
-
-                    SetAllOperationsToAborted(user, curResource);
-
-                    _logger.LogInformation(SepesEventId.SandboxDelete, "Study {0}, Sandbox {1}: Marking resource {2} for deletion", studyId, sandboxId, curResource.Id);
-                }
-
-                if (sandboxResourceGroup == null)
-                {
-                    throw new Exception($"Unable to find ResourceGroup record in DB for Sandbox {sandboxId}, StudyId: {studyId}.");
-                }
-
-                _logger.LogInformation(SepesEventId.SandboxDelete, $"Creating delete operation for resource group {sandboxResourceGroup.ResourceGroupName}");
-
-                var deleteOperation = new SandboxResourceOperation()
-                {
-                    CreatedBy = user.UserName,
-                    BatchId = Guid.NewGuid().ToString(),
-                    CreatedBySessionId = _requestIdService.GetRequestId(),
-                    OperationType = CloudResourceOperationType.DELETE,
-                    SandboxResourceId = sandboxResourceGroup.Id,
-                    Description = AzureResourceUtil.CreateDescriptionForResourceOperation(sandboxResourceGroup.ResourceType, CloudResourceOperationType.DELETE, sandboxResourceGroup.SandboxId) + ". (Delete of SandBox resource group and all resources within)",
-                    MaxTryCount = CloudResourceConstants.RESOURCE_MAX_TRY_COUNT
-                };
-
-                sandboxResourceGroup.Operations.Add(deleteOperation);
-
-                await _db.SaveChangesAsync();
-
-                _logger.LogInformation(SepesEventId.SandboxDelete, "Study {0}, Sandbox {1}: Queuing operation", studyId, sandboxId);
-
-                //Create queue item
-                var queueParentItem = new ProvisioningQueueParentDto();
-                queueParentItem.SandboxId = sandboxId;
-                queueParentItem.Description = $"Delete resources for Sandbox: {sandboxId}";
-                queueParentItem.Children.Add(new ProvisioningQueueChildDto() { SandboxResourceOperationId = deleteOperation.Id });
-                await _provisioningQueueService.SendMessageAsync(queueParentItem, visibilityTimeout: TimeSpan.FromSeconds(10));
-            }
-            else
-            {
-                _logger.LogCritical(SepesEventId.SandboxDelete, "Study {0}, Sandbox {1}: Unable to find any resources for Sandbox", studyId, sandboxId);
-                await _db.SaveChangesAsync();
-            }
-
-            _logger.LogInformation(SepesEventId.SandboxDelete, "Study {0}, Sandbox {1}: Done", studyId, sandboxId);
-        }
-
-        void SetAllOperationsToAborted(UserDto currentUser, SandboxResource resource)
-        {
-            foreach (var curOp in resource.Operations)
-            {
-                curOp.Status = CloudResourceOperationState.ABORTED;
-                curOp.Updated = DateTime.UtcNow;
-                curOp.UpdatedBy = currentUser.UserName;
-            }
-        }
-
-        public async Task ReScheduleSandboxCreation(int sandboxId)
-        {
-            var sandboxFromDb = await GetOrThrowAsync(sandboxId, UserOperation.Study_Crud_Sandbox, true);
-
-            var queueParentItem = new ProvisioningQueueParentDto();
-            queueParentItem.SandboxId = sandboxFromDb.Id;
-            queueParentItem.Description = $"Create basic resources for Sandbox (re-scheduled): {sandboxFromDb.Id}";
-
-            //Check state of sandbox resource creation: Resource group shold be success, rest should be not started or failed
-
-            var resourceGroupResource = sandboxFromDb.Resources.SingleOrDefault(r => r.ResourceType == AzureResourceType.ResourceGroup);
-
-            if (resourceGroupResource == null)
-            {
-                throw new NullReferenceException(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, "Could not locate database entry for ResourceGroup"));
-            }
-
-            var resourceGroupResourceOperation = resourceGroupResource.Operations.OrderByDescending(o => o.Created).FirstOrDefault();
-
-            if (resourceGroupResourceOperation == null)
-            {
-                throw new NullReferenceException(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, "Could not locate ANY database entry for ResourceGroupOperation"));
-            }
-            else if (resourceGroupResourceOperation.OperationType != CloudResourceOperationType.CREATE && resourceGroupResourceOperation.Status == CloudResourceOperationState.DONE_SUCCESSFUL)
-            {
-                throw new Exception(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, "Could not locate RELEVANT database entry for ResourceGroupOperation"));
-            }
-
-            var operations = new List<SandboxResourceOperation>();
-
-            foreach (var curResource in sandboxFromDb.Resources)
-            {
-                if (curResource.Id == resourceGroupResource.Id)
-                {
-                    //allready covered this above
-                    continue;
-                }
-
-                var relevantOperation = curResource.Operations.OrderByDescending(o => o.Created).FirstOrDefault();
-
-                if (relevantOperation == null)
-                {
-                    throw new NullReferenceException(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, "Could not locate ANY database entry for resource", curResource.Id));
-                }
-                else if (String.IsNullOrWhiteSpace(relevantOperation.Status) || relevantOperation.Status == CloudResourceOperationState.NEW || relevantOperation.Status == CloudResourceOperationState.IN_PROGRESS || relevantOperation.Status == CloudResourceOperationState.FAILED || relevantOperation.Status == CloudResourceOperationState.DONE_SUCCESSFUL)
-                {
-                    _logger.LogInformation(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, $"Re-queing item. Previous status was {relevantOperation.Status}", curResource.Id));
-                    relevantOperation.MaxTryCount += CloudResourceConstants.RESOURCE_MAX_TRY_COUNT; //Increase max try count               
-                    queueParentItem.Children.Add(new ProvisioningQueueChildDto() { SandboxResourceOperationId = relevantOperation.Id });
-                }
-                else
-                {
-                    throw new Exception(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, $"Could not locate RELEVANT database entry for resource", curResource.Id));
-                }
-            }
-
             await _db.SaveChangesAsync();
 
-            if (queueParentItem.Children.Count == 0)
-            {
-                throw new Exception(ReScheduleLogPrefix(sandboxFromDb.StudyId, sandboxId, $"Could not re-shedule creation. No relevant resource items found"));
-            }
-            else
-            {
-                await _provisioningQueueService.SendMessageAsync(queueParentItem);
-            }
-        }
+            await _sandboxCloudResourceService.HandleSandboxDeleteAsync(sandboxId);           
 
-        string ReScheduleLogPrefix(int studyId, int sandboxId, string logText, int resourceId = 0)
-        {
-            var logMessage = $"ReScheduleSandboxCreation | Study {studyId} | Sandbox {sandboxId}";
-
-            if (resourceId > 0)
-            {
-                logMessage += $" | Resource: {resourceId}";
-            }
-
-            logMessage += $" | {logText}";
-
-            return logMessage;
-        }
+            _logger.LogInformation(SepesEventId.SandboxDelete, "Study {0}, Sandbox {1}: Done", studyId, sandboxId);
+        }     
     }
 }
