+++ src/Sepes.RestApi/Controllers/SandboxResourceController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
+using Sepes.Infrastructure.Constants;
 using Sepes.Infrastructure.Service.Interface;
 using System.Threading.Tasks;
 
@@ -13,25 +14,35 @@ namespace Sepes.RestApi.Controller
     [Authorize]
     public class SandboxResourceController : ControllerBase
     {
-        readonly ISandboxService _sandboxService;
+        readonly ISandboxResourceService _sandboxResourceService;
+        readonly ISandboxCloudResourceService _sandboxCloudResourceService;
 
-        public SandboxResourceController(ISandboxService sandboxService)
+        public SandboxResourceController(ISandboxResourceService sandboxResourceService, ISandboxCloudResourceService sandboxCloudResourceService)
         {
-            _sandboxService = sandboxService;
+            _sandboxResourceService = sandboxResourceService;
+            _sandboxCloudResourceService = sandboxCloudResourceService;
         }       
 
         [HttpGet("sandboxes/{sandboxId}/resources")]
         public async Task<IActionResult> GetSandboxResources(int sandboxId)
         {
-            var sandboxes = await _sandboxService.GetSandboxResources(sandboxId);
+            var sandboxes = await _sandboxResourceService.GetSandboxResources(sandboxId);
             return new JsonResult(sandboxes);
         }
 
         [HttpPut("resources/{resourceId}/retry")]
         public async Task<IActionResult> RetryLastOperation(int resourceId)
         {
-            var resource = await _sandboxService.RetryLastOperation(resourceId);
+            var resource = await _sandboxCloudResourceService.RetryLastOperation(resourceId);
             return new JsonResult(resource);
-        }     
+        }
+
+        [HttpPut("sandboxes/{sandboxId}/retryCreate")]
+        [Authorize(Roles = AppRoles.Admin)]
+        public async Task<IActionResult> ReScheduleCreation(int sandboxId)
+        {
+            await _sandboxCloudResourceService.ReScheduleSandboxResourceCreation(sandboxId);
+            return new NoContentResult();
+        }
     }
 }
